services/chatbot/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-decouple==3.8
google-genai==0.3.0
pydantic==2.10.3
